Добавил поддержку сравнения epf- и erf-файлов (файлы 1С:Предприятия 8).

+++ diff-1c-77.py
@@ -42,11 +42,22 @@ def main():
         sys.path.append('C:\\Python34\\pycharm-debug-py3k.egg')
 
         import pydevd
-        pydevd.settrace(port=10050)  # todo
+        pydevd.settrace(port=10050)
 
+    exe1c = Path(get_setting('General', '1C'))
+    if not exe1c.exists():
+        raise Exception('Платформа не существует!')
+    ib = Path(get_setting('General', 'IB'))
+    if not ib.exists():
+        raise Exception('Сервисной информационной базы не существует!')
+    v8_reader = Path(get_setting('General', 'V8Reader'))
+    if not v8_reader.exists():
+        raise Exception('V8Reader не существует!')
     gcomp = Path(get_setting('General', 'GComp'))
+    if not gcomp.exists():
+        raise Exception('GComp не существует!')
 
-    # file1
+    # base
     base_path = Path(args.base)
     base_temp_path = Path(tempfile.mktemp(base_path.suffix))
     shutil.copyfile(str(base_path), str(base_temp_path))
@@ -55,21 +66,33 @@ def main():
     if not base_source_path.exists():
         base_source_path.mkdir(parents=True)
     else:
-        shutil.rmtree(str(base_source_path), ignore_errors=True)  # fixme
+        shutil.rmtree(str(base_source_path), ignore_errors=True)
 
     base_bat_path = Path(tempfile.mktemp('.bat'))
     with base_bat_path.open('w', encoding='cp866') as base_bat:
         base_bat.write('@echo off\n')
-        base_bat.write('"{}" -d -F "{}" -DD "{}"'.format(  # fixme
-            str(gcomp),
-            str(base_temp_path),
-            str(base_source_path)
-        ))
-    exit_code = subprocess.check_call(['cmd.exe', '/C', str(base_bat_path)])  # fixme
+        base_path_suffix_lower = base_path.suffix.lower()
+        if base_path_suffix_lower in ['.epf', '.erf']:
+            base_bat.write('"{}" /F"{}" /DisableStartupMessages /Execute"{}" {}'.format(
+                str(exe1c),
+                str(ib),
+                str(v8_reader),
+                '/C"decompile;pathtocf;{};pathout;{};shutdown;convert-mxl2txt;"'.format(
+                    str(base_temp_path),
+                    str(base_source_path)
+                )
+            ))
+        elif base_path_suffix_lower in ['.ert', '.md']:
+            base_bat.write('"{}" -d -F "{}" -DD "{}"'.format(
+                str(gcomp),
+                str(base_temp_path),
+                str(base_source_path)
+            ))
+    exit_code = subprocess.check_call(['cmd.exe', '/C', str(base_bat_path)])
     if not exit_code == 0:
-        raise Exception('Не удалось разобрать файл {}'.format(str(base_path)))  # fixme
+        raise Exception('Не удалось разобрать файл {}'.format(str(base_path)))
 
-    # file2
+    # mine
     mine_path = Path(args.mine)
     mine_temp_path = Path(tempfile.mktemp(mine_path.suffix))
     shutil.copyfile(str(mine_path), str(mine_temp_path))
@@ -78,19 +101,31 @@ def main():
     if not mine_source_path.exists():
         mine_source_path.mkdir(parents=True)
     else:
-        shutil.rmtree(str(mine_source_path), ignore_errors=True)  # fixme
+        shutil.rmtree(str(mine_source_path), ignore_errors=True)
 
     mine_bat_path = Path(tempfile.mktemp('.bat'))
     with mine_bat_path.open('w', encoding='cp866') as mine_bat:
         mine_bat.write('@echo off\n')
-        mine_bat.write('"{}" -d -F "{}" -DD "{}"'.format(  # fixme
-            str(gcomp),
-            str(mine_temp_path),
-            str(mine_source_path)
-        ))
-    exit_code = subprocess.check_call(['cmd.exe', '/C', str(mine_bat_path)])  # fixme
+        mine_path_suffix_lower = base_path.suffix.lower()
+        if mine_path_suffix_lower in ['.epf', '.erf']:
+            mine_bat.write('"{}" /F"{}" /DisableStartupMessages /Execute"{}" {}'.format(
+                str(exe1c),
+                str(ib),
+                str(v8_reader),
+                '/C"decompile;pathtocf;{};pathout;{};shutdown;convert-mxl2txt;"'.format(
+                    str(mine_temp_path),
+                    str(mine_source_path)
+                )
+            ))
+        elif mine_path_suffix_lower in ['.ert', '.md']:
+            mine_bat.write('"{}" -d -F "{}" -DD "{}"'.format(
+                str(gcomp),
+                str(mine_temp_path),
+                str(mine_source_path)
+            ))
+    exit_code = subprocess.check_call(['cmd.exe', '/C', str(mine_bat_path)])
     if not exit_code == 0:
-        raise Exception('Не удалось разобрать файл {}'.format(str(mine_path)))  # fixme
+        raise Exception('Не удалось разобрать файл {}'.format(str(mine_path)))
 
     tool_args = None
     if args.tool == 'KDiff3':
@@ -166,10 +201,10 @@ def main():
                 '--right_display_name:{}'.format(args.yname)
             ]
     if tool_args == None:
-        raise Exception('Не удалось сравнить файлы {} и {}'.format(str(base_path), str(mine_path)))  # fixme
+        raise Exception('Не удалось сравнить файлы {} и {}'.format(str(base_path), str(mine_path)))
     exit_code = subprocess.check_call(tool_args)
     if not exit_code == 0:
-        raise Exception('Не удалось сравнить файлы {} и {}'.format(str(base_path), str(mine_path)))  # fixme
+        raise Exception('Не удалось сравнить файлы {} и {}'.format(str(base_path), str(mine_path)))
 
 
 if __name__ == '__main__':
